all error slove this

+++ src/components/MyCart/MyCart.jsx
@@ -9,7 +9,7 @@ const MyCart = () => {
 
   useEffect(() => {
     fetch(
-      `https://technology-and-electronics-server-nadg6amp7.vercel.app/userData/${user?.email}`
+      `https://technology-and-electronics-server-beige.vercel.app/userData/${user?.email}`
     )
       .then((res) => res.json())
       //   .then((data) => setMyData(data));
@@ -25,29 +25,67 @@ const MyCart = () => {
   //
 
   const handleDeleted = (id) => {
-    //make sure user is confirmed to deled
-    fetch(
-      `https://technology-and-electronics-server-nadg6amp7.vercel.app/userData/${id}`,
-      {
-        method: "DELETE",
-      }
-    )
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(data);
-        if (data.deletedCount > 0) {
-          console.log("Deleted successfully");
-          Swal.fire(
-            "Deleted!",
-            "Your Product has been deleted successfully",
-            "success"
-          );
+    //make sure user is confirmed to deleted
 
-          // remove the user from the ui
-          const remainingUsers = myData.filter((user) => user?._id !== id);
-          setMyData(remainingUsers);
-        }
-      });
+    // fetch(
+    //   `https://technology-and-electronics-server-beige.vercel.app/userData/${id}`,
+    //   {
+    //     method: "DELETE",
+    //   }
+    // )
+    //   .then((res) => res.json())
+    //   .then((data) => {
+    //     console.log(data);
+    //     if (data.deletedCount > 0) {
+    //       console.log("Deleted successfully");
+    //       Swal.fire(
+    //         "Deleted!",
+    //         "Your Product has been deleted successfully",
+    //         "success"
+    //       );
+
+    //       // remove the user from the ui
+    //       const remainingUsers = myData.filter((user) => user?._id !== id);
+    //       setMyData(remainingUsers);
+    //     }
+    //   });
+
+    // ----------------------------
+
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(
+          `https://technology-and-electronics-server-beige.vercel.app/userData/${id}`,
+          {
+            method: "DELETE",
+          }
+        )
+          .then((res) => res.json())
+          .then((data) => {
+            console.log(data);
+
+            if (data.deletedCount > 0) {
+              Swal.fire(
+                "Deleted!",
+                "Your Product has been Deleted.",
+                "success"
+              );
+
+              // remove the user from the ui
+              const remainingUsers = myData.filter((user) => user?._id !== id);
+              setMyData(remainingUsers);
+            }
+          });
+      }
+    });
   };
 
   return (
